ESC-339 move syntax classes into dedicated package
