+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/services/Journey.scala
@@ -20,7 +20,7 @@ import play.api.Logger
 import play.api.libs.json.{Json, OFormat}
 import play.api.mvc.Results.Redirect
 import play.api.mvc.{AnyContent, Request, Result}
-import uk.gov.hmrc.eusubsidycompliancefrontend.util.FutureSyntax.FutureOps
+import uk.gov.hmrc.eusubsidycompliancefrontend.syntax.FutureSyntax.FutureOps
 
 import scala.concurrent.Future
