+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/syntax/TaxYearSyntax.scala
@@ -14,14 +14,20 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.eusubsidycompliancefrontend.util
+package uk.gov.hmrc.eusubsidycompliancefrontend.syntax
 
-import scala.concurrent.Future
+import uk.gov.hmrc.eusubsidycompliancefrontend.util.TaxYearHelpers
 
-object FutureSyntax {
+import java.time.LocalDate
 
-  implicit class FutureOps[A](val a: A) extends AnyVal {
-    def toFuture: Future[A] = Future.successful(a)
+object TaxYearSyntax {
+
+  implicit class LocalDateTaxYearOps(val d: LocalDate) extends AnyVal {
+    def toTaxYearStart: LocalDate = TaxYearHelpers.taxYearStartForDate(d)
+
+    def toTaxYearEnd: LocalDate = TaxYearHelpers.taxYearEndForDate(d)
+
+    def toEarliestTaxYearStart: LocalDate = TaxYearHelpers.earliestAllowedDate(d)
   }
 
 }
